+++ flutter_local_notifications/lib/flutter_local_notifications.dart
@@ -9,6 +9,7 @@ export 'src/initialization_settings.dart';
 export 'src/notification_details.dart';
 export 'src/platform_flutter_local_notifications.dart'
     hide MethodChannelFlutterLocalNotificationsPlugin;
+export 'src/platform_specifics/android/active_notification.dart';
 export 'src/platform_specifics/android/bitmap.dart';
 export 'src/platform_specifics/android/enums.dart'
     hide AndroidBitmapSource, AndroidIconSource, AndroidNotificationSoundSource;